create `useCurrentProject.tsx` and move related hooks
