+++ react/src/hooks/index.tsx
@@ -2,7 +2,7 @@ import { useEventNotStable } from './useEventNotStable';
 import { useAtomValue, useSetAtom } from 'jotai';
 import { atomWithDefault } from 'jotai/utils';
 import _ from 'lodash';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useQuery } from 'react-query';
 import { NavigateOptions, To, useNavigate } from 'react-router-dom';
 
@@ -64,49 +64,6 @@ export const useUpdatableState = (initialValue: string) => {
 export const useCurrentDomainValue = () => {
   const baiClient = useSuspendedBackendaiClient();
   return baiClient._config.domainName;
-};
-
-const currentProjectAtom = atomWithDefault(() => {
-  return {
-    // @ts-ignore
-    name: globalThis?.backendaiclient?.current_group,
-    // @ts-ignore
-    id: globalThis?.backendaiclient?.current_group_id(),
-  };
-});
-
-export const useCurrentProjectValue = () => {
-  useSuspendedBackendaiClient();
-  return useAtomValue(currentProjectAtom);
-};
-
-export const useSetCurrentProject = () => {
-  const set = useSetAtom(currentProjectAtom);
-  const baiClient = useSuspendedBackendaiClient();
-  return useCallback(
-    ({
-      projectName,
-      projectId,
-    }: {
-      projectName: string;
-      projectId: string;
-    }) => {
-      set({
-        name: projectName,
-        id: projectId,
-      });
-
-      // To sync with baiClient
-      baiClient.current_group = projectName;
-      // @ts-ignore
-      globalThis.backendaiutils._writeRecentProjectGroup(projectName);
-      const event: CustomEvent = new CustomEvent('backend-ai-group-changed', {
-        detail: projectName,
-      });
-      document.dispatchEvent(event);
-    },
-    [set, baiClient],
-  );
 };
 
 export const useAnonymousBackendaiClient = ({
